Move generic dependency registration to Foundation
