+++ account-management/src/Application/ApplicationConfiguration.cs
@@ -1,9 +1,6 @@
 using System.Reflection;
-using MediatR;
 using Microsoft.Extensions.DependencyInjection;
 using PlatformPlatform.AccountManagement.Application.Tenants.Dtos;
-using PlatformPlatform.Foundation.Application.DomainEvents;
-using PlatformPlatform.Foundation.Application.Persistence;
 
 namespace PlatformPlatform.AccountManagement.Application;
 
@@ -18,9 +15,6 @@ public static class ApplicationConfiguration
     public static IServiceCollection AddApplicationServices(this IServiceCollection services)
     {
         services.AddMediatR(configuration => configuration.RegisterServicesFromAssemblies(Assembly));
-
-        services.AddTransient(typeof(IPipelineBehavior<,>), typeof(UnitOfWorkPipelineBehavior<,>));
-        services.AddTransient(typeof(IPipelineBehavior<,>), typeof(PublishDomainEventsPipelineBehavior<,>));
 
         ConfigureMappings();
