+++ account-management/src/Infrastructure/Infrastructure.csproj
@@ -12,13 +12,4 @@
         <ProjectReference Include="..\Application\Application.csproj"/>
     </ItemGroup>
 
-    <ItemGroup>
-        <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="7.0.4"/>
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="7.0.4">
-            <PrivateAssets>all</PrivateAssets>
-            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-        </PackageReference>
-        <PackageReference Include="Scrutor" Version="4.2.2"/>
-    </ItemGroup>
-
 </Project>
