+++ account-management/src/WebApi/Program.cs
@@ -2,12 +2,14 @@ using PlatformPlatform.AccountManagement.Application;
 using PlatformPlatform.AccountManagement.Infrastructure;
 using PlatformPlatform.AccountManagement.WebApi;
 using PlatformPlatform.AccountManagement.WebApi.Endpoints;
+using PlatformPlatform.Foundation;
 using PlatformPlatform.Foundation.WebApi;
 
 var builder = WebApplication.CreateBuilder(args);
 
 // Configure services for the Application, Infrastructure, and WebApi layers.
 builder.Services
+    .AddFoundationServices()
     .AddApplicationServices()
     .AddInfrastructureServices(builder.Configuration)
     .AddWebApiServices();
